vsc_ast: move parsing code away from vsc_ast to vsc_parser
Moving all code related to YAML processing to the vsc_parser.py
from vsc_ast.py to keep in vsc_ast module only AST nodes.
