+++ vsc/model/vsc_ast.py
@@ -3,8 +3,7 @@
 # SPDX-License-Identifier: MPL-2.0
 
 from dataclasses import dataclass, field
-from typing import List, Optional, Dict, Any
-import yaml, dacite
+from typing import List, Optional
 
 
 @dataclass
@@ -118,38 +117,3 @@ class Namespace:
 class AST(Namespace):
     pass
 
-
-def read_yaml_file(filename) -> str:
-    """
-    Tries to read a file which contains yaml into a string
-    TODO: can have performance implications when file size is big. We have to consider incremental yaml processing.
-    :param filename:
-    :return: file contents as string
-    """
-    with open(filename, 'r') as yaml_file:
-        return yaml_file.read()
-
-
-def parse_yaml_file(yaml_string: str) -> Dict[Any, Any]:
-    """
-    Tries to parse yaml into a python dictionary
-    :param yaml_string: String containing text in YAML format
-    :return: Dictionary
-    """
-    return yaml.safe_load(yaml_string)
-
-
-def read_ast_from_yaml_file(filename: str) -> AST:
-    """
-    Reads a yaml file and returns AST
-    :param filename: path to a yaml file
-    :return: abstract syntax tree (vehicle service catalog)
-    """
-
-    yaml_string = read_yaml_file(filename)
-
-    yaml_dict = parse_yaml_file(yaml_string)
-
-    ast = dacite.from_dict(data_class=AST, data=yaml_dict)
-
-    return ast
